Initial fix of test suite
